Moved Key to its own mod, split document.rs
