+++ crates/bonsaidb-core/src/key.rs
